feat : finished likes, deslikes and likes count
