+++ src/components/Posts.js
@@ -7,10 +7,12 @@ export default function Posts() {
 
     const [posts, setPosts] = useState([])
     const [message, setMessage] = useState("Loading...")
+    const [callApi, setCallApi] = useState(true)
+    const [render, setRender] = useState(0)
 
     useEffect(async ()=> {
         try {
-            const response = await getPostsData() //passar config
+            const response = await getPostsData({userId: 1}) //userId de teste
             
             if (response.data.length === 0) {
                 setMessage("There are no posts yet")
@@ -21,17 +23,17 @@ export default function Posts() {
             setMessage("An error occured while trying to fetch the posts, please refresh the page")
             console.log(error)
         }
-    }, [])
+    }, [callApi])
 
 
     return (
         <Container>
             {posts.length > 0 ? 
             posts.map((value, index) => <Post key = {index} 
-            username = {value.username} picture_url = {value.picture_url} 
-            body = {value.body} post_url = {value.post_url} metadata = {value.metadata}/>) :
-            <LoadMessage>{message}</LoadMessage>
-        }
+            username = {value.username} picture_url = {value.picture_url} postId = {value.id}
+            body = {value.body} post_url = {value.post_url} metadata = {value.metadata} 
+            liked = {value.liked} likesCount = {value.likesCount} callApi = {callApi} setCallApi = {setCallApi}/>) :
+            <LoadMessage>{message}</LoadMessage>}
         </Container>
     )
 }
@@ -48,7 +50,7 @@ const LoadMessage = styled.div`
     width: 200px;
     margin: 340px auto;
     color: white;
-    font-family: Arial, Helvetica, sans-serif; //trocar pra Lato
+    font-family: 'Lato', sans-serif;
     font-size: 19px;
     font-weight: 400;
     text-align: center;
